apps/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery[redis]==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Database drivers
psycopg2-binary==2.9.9

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# File processing libraries
pypdf2==3.0.1
pillow==10.1.0
python-multipart==0.0.6

# HTTP client
httpx==0.25.2

# Utilities
python-jose[cryptography]==3.3.0
uuid==1.30
